Rename @/tests/mock -> @/tests/data
