+++ components/bank/components/__tests__/sendBox.test.tsx
@@ -3,7 +3,7 @@ import { afterEach, describe, expect, jest, mock, test } from 'bun:test';
 import React from 'react';
 
 import SendBox from '@/components/bank/components/sendBox';
-import { mockBalances } from '@/tests/mock';
+import { mockBalances } from '@/tests/data';
 import { renderWithChainProvider } from '@/tests/render';
 
 // Mock next/router
