+++ components/factory/forms/__tests__/MintForm.test.tsx
@@ -3,7 +3,7 @@ import { afterEach, describe, expect, jest, mock, test } from 'bun:test';
 import React from 'react';
 
 import MintForm from '@/components/factory/forms/MintForm';
-import { mockDenomMeta1, mockFakeMfxDenom, mockMfxDenom } from '@/tests/mock';
+import { mockDenomMeta1, mockFakeMfxDenom, mockMfxDenom } from '@/tests/data';
 import { renderWithChainProvider } from '@/tests/render';
 
 // Mock next/router
