+++ components/groups/components/__tests__/myGroups.test.tsx
@@ -3,7 +3,7 @@ import { afterEach, beforeEach, describe, expect, jest, mock, test } from 'bun:t
 import React from 'react';
 
 import YourGroups from '@/components/groups/components/myGroups';
-import { mockGroup, mockGroup2 } from '@/tests/mock';
+import { mockGroup, mockGroup2 } from '@/tests/data';
 import { renderWithChainProvider } from '@/tests/render';
 
 // Mock next/router
